Fix assembly circular dependency: move data-typed events to Game.Data
RecipeData and PillData live in Game.Data, which already references
Game.Core, making it impossible for Game.Core to reference them back.

- Create GameDataEvents.cs in Game.Data with crafting and pill events
- Remove those events from GameEvents.cs (Game.Core) along with the
  now-unneeded `using CultivationGame.Data` import
- Update CraftingSystem, CraftingUI, PlayerInventory, PillBuffSystem
  to use GameDataEvents instead of GameEvents for these events
