+++ Assets/_Project/Scripts/Core/GameEvents.cs
@@ -1,5 +1,3 @@
-using CultivationGame.Data;
-
 namespace CultivationGame.Core
 {
     public static class GameEvents
@@ -57,32 +55,5 @@ namespace CultivationGame.Core
         public static event MeditationBonusApplied OnMeditationBonusApplied;
         public static void RaiseMeditationBonusApplied(float multiplier)
             => OnMeditationBonusApplied?.Invoke(multiplier);
-
-        // --- Crafting ---
-        public delegate void CraftingStarted(RecipeData recipe);
-        public static event CraftingStarted OnCraftingStarted;
-        public static void RaiseCraftingStarted(RecipeData recipe)
-            => OnCraftingStarted?.Invoke(recipe);
-
-        public delegate void CraftingCompleted(RecipeData recipe);
-        public static event CraftingCompleted OnCraftingCompleted;
-        public static void RaiseCraftingCompleted(RecipeData recipe)
-            => OnCraftingCompleted?.Invoke(recipe);
-
-        public delegate void CraftingFailed(RecipeData recipe);
-        public static event CraftingFailed OnCraftingFailed;
-        public static void RaiseCraftingFailed(RecipeData recipe)
-            => OnCraftingFailed?.Invoke(recipe);
-
-        // --- Pills ---
-        public delegate void PillConsumed(PillData pill);
-        public static event PillConsumed OnPillConsumed;
-        public static void RaisePillConsumed(PillData pill)
-            => OnPillConsumed?.Invoke(pill);
-
-        public delegate void PillEffectsApplied(PillData pill, float effectiveness);
-        public static event PillEffectsApplied OnPillEffectsApplied;
-        public static void RaisePillEffectsApplied(PillData pill, float effectiveness)
-            => OnPillEffectsApplied?.Invoke(pill, effectiveness);
     }
 }
